WIP: Moving transaction manager to src
