+++ example/account_handling/AccountHandling.cpp
@@ -22,7 +22,7 @@
 #include <boost/program_options.hpp>
 #include <boost/format.hpp>
 #include <boost/asio.hpp>
-#include "TransactionManager.hpp"
+#include "account/TransactionManager.hpp"
 #include "blockchain/impl/common.hpp"
 #include "blockchain/impl/key_value_block_header_repository.hpp"
 #include "blockchain/impl/key_value_block_storage.hpp"
